integration_2022_wk49

See merge request oai/openairinterface5g!1851

!1714 first-fix-rrc-mac-interface
!1755 DL/UL channel estimation improvement
!1788 NR RRC harmonization of CSI reporting
!1848 Remove old RNTI at MAC layer when Msg3 RRCSetupRequest carries ng-5G-S-TMSI-Part1
deactivate SRS in SA tests: creates RT problems
!1809 Doppler Effect
!1831 Nr ue fix asn1 memleak
!1842 Parallelize ASN.1 generation and build
!1853 Correct ue_ip driver module for RHEL9
!1850 Fix T1 offload test
!1461 Optimize LTE TM2 modulation (doesn't work well but it is already the case)
!1806 Repair doxygen documentation generation
!1856 hotfix for nr_rrc_data_req for the unfortunate SDU size 255

+++ openair1/PHY/CODING/crc_byte.c
@@ -298,11 +298,11 @@ crc8 (unsigned char * inptr, int bitlen)
   resbit = (bitlen % 8);
 
   while (octetlen-- > 0) {
-    crc = crc8Table[(*inptr++) ^ (crc >> 24)] << 24;
+    crc = ((unsigned int)crc8Table[(*inptr++) ^ (crc >> 24)]) << 24;
   }
 
   if (resbit > 0)
-    crc = (crc << resbit) ^ (crc8Table[((*inptr) >> (8 - resbit)) ^ (crc >> (32 - resbit))] << 24);
+    crc = (crc << resbit) ^ ((unsigned int)(crc8Table[((*inptr) >> (8 - resbit)) ^ (crc >> (32 - resbit))]) << 24);
 
   return crc;
 }
